Move CreateDenoiseMaskInvocation to its own file. No functional changes.
